refactor(types): 将基本类型改为其包装类型以支持空值
- 将多个 int、long、boolean 类型属性及变量修改为 Integer、Long、Boolean 等包装类型
- 修改方法返回类型与参数类型以匹配包装类型需求
- 解决部分场景中基本类型无法表示 null 的限制，提升数据处理灵活性
- 统一代码中类似类型定义，增强一致性与可维护性
- 调整分布式锁等核心逻辑中数值类型为包装类型，避免潜在空指针风险
- 导入相关 DTO、实体、服务和工具类均对应改造类型规范

+++ src/main/java/com/xrq/xxq/common/PageQuery.java
@@ -28,12 +28,12 @@ public class PageQuery {
     private Integer pageSize;
 
     /** 规范化页码：null 或小于 1 时取默认值。 */
-    public int resolvedPage() {
+    public Integer resolvedPage() {
         return (page == null || page < 1) ? DEFAULT_PAGE : page;
     }
 
     /** 规范化每页条数：null 或小于 1 取默认值，超过上限则截断。 */
-    public int resolvedSize() {
+    public Integer resolvedSize() {
         if (pageSize == null || pageSize < 1) {
             return DEFAULT_SIZE;
         }
@@ -42,6 +42,6 @@ public class PageQuery {
 
     /** 构造 MyBatis Plus 分页对象（含 total 查询）。 */
     public <T> Page<T> toPage() {
-        return new Page<>(resolvedPage(), resolvedSize());
+        return new Page<>((long) resolvedPage(), (long) resolvedSize());
     }
 }
